+++ source/main/modules/fsal/fsal.js
@@ -47,10 +47,12 @@ module.exports = class FSAL extends EventEmitter {
 
     // The following actions can be run on the file tree
     this._actions = {
+      // TODO: Doesn't emit currently
       'sort': async (src, target, options) => {
         // NOTE: Does not generate watchdog events
         return FSALDir.sort(src, options)
       },
+      // TODO: Doesn't emit currently
       'create-file': async (src, target, options) => {
         // This action needs the cache because it'll parse a file
         // NOTE: Generates an add-event
@@ -60,6 +62,7 @@ module.exports = class FSAL extends EventEmitter {
         })
         return FSALDir.createFile(src, options, this._cache)
       },
+      // TODO: Doesn't emit currently
       'duplicate-file': async (src, target, options) => {
         // Duplicating a file is basically the same as creating, only with
         // passing the content of an existing file to the createFile
@@ -86,7 +89,7 @@ module.exports = class FSAL extends EventEmitter {
           'path': path.join(path.dirname(src.path), options.name)
         })
 
-        await FSALFile.rename(src, options)
+        await FSALFile.rename(src, this._cache, options)
         // Now we need to re-sort the parent directory
         if (!isRoot) await FSALDir.sort(src.parent) // Omit sorting
 
@@ -96,6 +99,7 @@ module.exports = class FSAL extends EventEmitter {
         if (isOpenFile) this.emit('fsal-state-changed', 'openFiles')
         return true
       },
+      // TODO: Doesn't emit currently
       'remove-file': async (src, target, options) => {
         // NOTE: Generates 1x unlink
         // First remove the file
@@ -108,12 +112,9 @@ module.exports = class FSAL extends EventEmitter {
         this.closeFile(src) // Does nothing if the file is not open
       },
       'save-file': async (src, target, options) => {
-        // NOTE: Generates 1x chance
-        this._watchdog.ignoreEvents({
-          'event': 'change',
-          'path': src.path
-        })
-        await FSALFile.save(src, options)
+        // NOTE: Generates 1x change
+        this._watchdog.ignoreEvents({ 'event': 'change', 'path': src.path })
+        await FSALFile.save(src, this._cache, options)
         // Notify that a file has saved, which strictly speaking does not
         // modify the openFiles array, but does change the modification flag.
         this.emit('fsal-state-changed', 'fileSaved')
@@ -125,15 +126,18 @@ module.exports = class FSAL extends EventEmitter {
         return FSALFile.search(src, options)
       },
       // Creates a project in a dir
+      // TODO: Doesn't emit currently
       'create-project': async (src, target, options) => {
         // NOTE: Generates no events as dotfiles are not watched
         await FSALDir.makeProject(src, options)
       },
+      // TODO: Doesn't emit currently
       'update-project': async (src, target, options) => {
         // NOTE: Generates no events as dotfiles are not watched
         // Updates the project properties on a directory.
         await FSALDir.updateProjectProperties(src, options)
       },
+      // TODO: Doesn't emit currently
       'remove-project': async (src, target, options) => {
         // NOTE: Generates no events as dotfiles are not watched
         await FSALDir.removeProject(src)
@@ -146,6 +150,12 @@ module.exports = class FSAL extends EventEmitter {
           'path': path.join(src.path, options.name)
         })
         await FSALDir.create(src, options, this._cache)
+
+        // Notify the event listeners
+        this.emit('fsal-state-changed', 'directory', {
+          'oldHash': src.hash,
+          'newHash': src.hash
+        })
       },
       'rename-directory': async (src, target, options) => {
         // We are probably going to need that code from the move action
@@ -153,42 +163,28 @@ module.exports = class FSAL extends EventEmitter {
         let openDirUpdateNeeded = false
         let newOpenDirHash
         let newFileHashes = []
-        if (src.type === 'directory') {
-          // A directory is being moved, so check the open files if something
-          // needs to change concerning them.
-          for (let file of this._state.openFiles) {
-            let found = this.findFile(file.hash, src)
-            if (found) {
-              // The file is in there, so we need to update the open files
-              openFilesUpdateNeeded = true
-              // Exchange the old directory path for the new one and compute
-              // its new hash
-              let newHash = hash(file.path.replace(src.parent.path, target.path))
-              newFileHashes.push(newHash)
-            } else {
-              // Nothing really to do
-              newFileHashes.push(file.hash)
-            }
-          }
-        } else if (src.type === 'file') {
-          if (this._state.openFiles.includes(src)) {
-            // The source is an open file, we need to account for that.
+        // A directory is being moved, so check the open files if something
+        // needs to change concerning them.
+        for (let file of this._state.openFiles) {
+          let found = this.findFile(file.hash, src)
+          if (found) {
+            // The file is in there, so we need to update the open files
             openFilesUpdateNeeded = true
-            let newHash = hash(src.path.replace(src.parent.path, target.path))
+            // Exchange the old directory path for the new one and compute
+            // its new hash
+            let newHash = hash(file.path.replace(path.dirname(src.path), target.path))
             newFileHashes.push(newHash)
-            newFileHashes = this._state.openFiles.map(e => e.hash)
-            newFileHashes.splice(newFileHashes.indexOf(src.hash), 1, newHash)
+          } else {
+            // Nothing really to do
+            newFileHashes.push(file.hash)
           }
         }
 
-        if (
-          src.type === 'directory' &&
-          (src === this._state.openDirectory ||
-          this.findDir(this._state.openDirectory.hash, src))
-        ) {
+        if (src === this._state.openDirectory ||
+          this.findDir(this._state.openDirectory.hash, src)) {
           // Compute the new hash and indicate an update is necessary
           openDirUpdateNeeded = true
-          newOpenDirHash = hash(this._state.openDirectory.path.replace(src.parent.path, target.path))
+          newOpenDirHash = hash(this._state.openDirectory.path.replace(path.dirname(src.path), target.path))
         }
 
         // Now that we have prepared potential updates,
@@ -217,17 +213,25 @@ module.exports = class FSAL extends EventEmitter {
 
         await FSALDir.rename(src, options, this._cache)
 
-        // Update the parent
-        this.emit('fsal-state-changed', 'directory', {
-          'oldHash': src.parent.hash,
-          'newHash': src.parent.hash
-        })
+        if (!src.parent) {
+          this.emit('fsal-state-changed', 'filetree')
+        } else {
+          // Update the parent
+          this.emit('fsal-state-changed', 'directory', {
+            'oldHash': src.parent.hash,
+            'newHash': src.parent.hash
+          })
+        }
+
+        // Make sure we have a correct set of open files
+        this._refetchOpenFiles()
 
         // Afterwards, let's see if we have to change something. These
         // functions will notify the application respectively.
         if (openFilesUpdateNeeded) this.setOpenFiles(newFileHashes)
         if (openDirUpdateNeeded) this.setOpenDirectory(this.findDir(newOpenDirHash))
       },
+      // TODO: Doesn't emit currently
       'remove-directory': async (src, target, options) => {
         // NOTE: Generates 1x unlink for each child + src!
         let arr = objectToArray(src, 'children').map(e => {
@@ -240,6 +244,7 @@ module.exports = class FSAL extends EventEmitter {
         this._watchdog.ignoreEvents(arr)
         await FSALDir.remove(src)
       },
+      // TODO: Doesn't emit currently (at least not consistently)
       'move': async (src, target, options) => {
         // NOTE: Generates 1x unlink, 1x add for each child, src and on the target!
         let openFilesUpdateNeeded = false
@@ -311,7 +316,7 @@ module.exports = class FSAL extends EventEmitter {
         // Now update both the source's parent and the target
         this.emit('fsal-state-changed', 'directory', {
           // We cannot move roots, so the source WILL have a parent
-          'oldHash': src.parent.hash,
+          'oldHash': src.parent.hash, // NOTE that src still points to the old location
           'newHash': src.parent.hash
         })
         this.emit('fsal-state-changed', 'directory', {
@@ -436,8 +441,6 @@ module.exports = class FSAL extends EventEmitter {
         let index = descriptor.parent.children.indexOf(descriptor)
         descriptor.parent.children.splice(index, 1, newdir)
       }
-      // Make sure to pull potential new openFiles from the filetree
-      this._refetchOpenFiles()
       isDirectoryUpdateNeeded = true
       directoryToUpdate = descriptor
     } else if (isRoot && event === 'unlinkDir') {
@@ -535,6 +538,11 @@ module.exports = class FSAL extends EventEmitter {
         }
       }
     } // END isOpenDir
+
+    // Make sure to pull potential new openFiles from the filetree. There is a
+    // variety of events that might change that list. We'll do this check here
+    // after everything that might have changed has changed for good.
+    this._refetchOpenFiles()
 
     // Finally, trigger all necessary events
     if (isDirectoryUpdateNeeded) {
@@ -987,6 +995,11 @@ module.exports = class FSAL extends EventEmitter {
   }
 
   getOpenDirectory () { return this._state.openDirectory }
+
+  /**
+   * Clears the cache
+   */
+  clearCache () { return this._cache.clearCache() }
 
   /**
    * Runs an action on the file tree